+++ src/App.tsx
@@ -1,14 +1,13 @@
-import axios from 'axios';
-import React, { useEffect } from 'react';
-import { useSelector } from 'react-redux';
+import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import Subway from './components/@common/Icon/Subway';
 import Header from './components/@shared/Header/Header';
 import Main from './components/@shared/Main/Main';
 import Navigation from './components/@shared/Navigation/Navigation';
-import { API_INFO } from './constants/api';
 import { APP_TITLE, COMMON_NAV_LIST, PAGE_INFO } from './constants/appInfo';
 import PALETTE from './constants/palette';
+import useAPIBaseURL from './hooks/useAPIBaseURL';
+import useLoginManager from './hooks/useLoginManager';
 import Home from './pages/Home/Home';
 import Lines from './pages/Lines/Lines';
 import Login from './pages/Login/Login';
@@ -16,23 +15,10 @@ import Sections from './pages/Sections/Sections';
 import Signup from './pages/Signup/Signup';
 import Stations from './pages/Stations/Stations';
 import SubwayMap from './pages/SubwayMap/SubwayMap';
-import { RootState } from './redux/store';
-import { getApiOwner, getBearerToken } from './storage/service';
-
-axios.defaults.baseURL = API_INFO[getApiOwner()].endPoint;
-axios.defaults.headers.common['Authorization'] = getBearerToken();
 
 const App = (): JSX.Element => {
-  const apiOwner = useSelector((state: RootState) => state.api.owner);
-  const isLogin = useSelector((state: RootState) => state.login.isLogin);
-
-  useEffect(() => {
-    axios.defaults.baseURL = API_INFO[apiOwner].endPoint;
-  }, [apiOwner]);
-
-  useEffect(() => {
-    axios.defaults.headers.common['Authorization'] = getBearerToken();
-  }, [isLogin]);
+  useAPIBaseURL();
+  useLoginManager();
 
   return (
     <>
